% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_list_crs.R
\name{list_crs}
\alias{list_crs}
\title{Generate a list of EML-compliant coordinate reference system (CRS)
names}
\usage{
list_crs()
}
\value{
A character vector of EML-compliant coordinate reference system
names. These values are valid inputs to the \code{coord_sys} parameter of
\code{\link{create_vector_shape}} and \code{\link{package_vector_shape}}.
}
\description{
list_crs is a helper function that provides convenient access
to a list of EML-compliant coordinate references systems (CRSs) by querying
the authoritative NCEAS EML schema directly.
}
\examples{
\dontrun{

# retrieve the full list of EML-compliant CRS names
crs_names <- capemlVector::list_crs()

# search for UTM zone options
crs_names[grepl("UTM", crs_names)]

}

}
